restructurize files, refactoring, connected sentry
