+++ app.py
@@ -9,6 +9,8 @@ from flask_talisman import Talisman
 from middlewares import CustomSessionInterface
 from flask_marshmallow import Marshmallow
 # from flask_wtf.csrf import CSRFProtect
+import sentry_sdk
+from sentry_sdk.integrations.flask import FlaskIntegration
 
 
 app = Flask(__name__)
@@ -18,6 +20,18 @@ if env:
     app.config.from_object(environments[env.get('ENVIRONMENT')])
 else:
     app.config.from_object(environments['Production'])
+
+sentry_sdk.init(
+    dsn=app.config.get('SENTRY_URI'),
+    integrations=[
+        FlaskIntegration(),
+    ],
+
+    # Set traces_sample_rate to 1.0 to capture 100%
+    # of transactions for performance monitoring.
+    # We recommend adjusting this value in production.
+    traces_sample_rate=1.0
+)
 
 # csrf = CSRFProtect(app)
 jwt = JWTManager(app)
