+++ requirements.txt
@@ -1,6 +1,8 @@
 alembic==1.9.1
 aniso8601==9.0.1
 bidict==0.22.0
+blinker==1.5
+certifi==2022.12.7
 click==8.1.3
 dnspython==1.16.0
 Flask==2.2.2
@@ -34,8 +36,10 @@ python-dotenv==0.21.0
 python-engineio==4.3.4
 python-socketio==5.7.2
 pytz==2022.7
+sentry-sdk==1.13.0
 six==1.16.0
 SQLAlchemy==1.4.45
+urllib3==1.26.14
 Werkzeug==2.2.2
 WTForms==3.0.1
 zope.event==4.6
