--- static/scripts/button_to_top.js
@@ -1,13 +1,0 @@
-$(function() {
-	$(window).scroll(function() {
-		if($(this).scrollTop() > 250) {
-			$('#toTop').fadeIn();
-		} else {
-			$('#toTop').fadeOut();
-		}
-	});
-	$('#toTop').click(function() {
-		$('body,html').animate({scrollTop:0}, 800);
-	});
-	
-});
