--- templates/admin/admin_index.html
@@ -1,106 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-    <link rel="stylesheet" href="./static/css/style.css" type="text/css"> 
-    <style type="text/css" rel="stylesheet" media="all">
-      /* Base ------------------------------ */
-      body {
-        width: 100% !important;
-        height: 100%;
-        margin: 0;
-        line-height: 1.4;
-        background-color: #F5F7F9;
-        color: #839197;
-        -webkit-text-size-adjust: none;
-      }
-      a {
-        text-decoration: none !important;
-      }
-    </style>
-    <title>Hello, world!</title>
-  </head>
-  <body>
-    <div class='container-fluid'>
-      <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="#">Admin</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
-          <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNavDropdown">
-          <ul class="navbar-nav" style='display: none;'>
-            <li class="nav-item">
-              <a class="nav-link" href="#">Users <span class="sr-only">(current)</span></a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="#">Posts</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="#">Categories</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" href="#">Tags</a>
-            </li>
-            <li class="nav-item dropdown">
-              <a class="nav-link dropdown-toggle" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                Dropdown link
-              </a>
-              <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                <a class="dropdown-item" href="#">...</a>
-                <a class="dropdown-item" href="#">...</a>
-                <a class="dropdown-item" href="#">...</a>
-              </div>
-            </li>
-          </ul>
-          <form class="form-inline">
-            <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
-            <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
-          </form>
-        </div>
-      </nav>
-      <div class='d-lg-flex' style='margin-top: 10px;'>
-        <div class='col shadow p-3 mb-5 bg-white rounded mr-2'>
-          <a href='#'>
-            <b>Users</b>
-          </a>
-          <a href='#' style='float: right;'>+</a>
-          <hr/>
-          <div>first user</div>
-        </div>
-        <div class='col shadow p-3 mb-5 bg-white rounded mr-2'>
-          <a href='#'>
-          <b>Posts</b>
-          </a>
-          <a href='#' style='float: right;'>+</a>
-          <hr/>
-          <div>first post</div>
-        </div>
-        <div class='col shadow p-3 mb-5 bg-white rounded mr-2'>
-          <a href='#'>
-          <b>Category</b>
-          </a>
-          <a href='#' style='float: right;'>+</a>
-          <hr/>
-          <div>first category</div>
-        </div>
-        <div class='col shadow p-3 mb-5 bg-white rounded mr-2'>
-          <a href='#'>
-          <b>Tags</b>
-          </a>
-          <a href='#' style='float: right;'>+</a>
-          <hr/>
-          <div>first tag</div>
-        </div>
-      </div>
-      </div>
-
-    <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
-  </body>
-</html>
