--- templates/admin/knots.html
@@ -1,14 +1,0 @@
-{% extends 'admin/master.html' %}
-
-{% block body %}
-	{% for knot in knots.items %}
-        <div class='post_container'>
-            <div class='post_paragraph'>
-                <!-- <a class='post_title' href="{{ url_for('posts.post_content', slug=post.slug) }}"> {{ post.title }} </a> -->
-                <div class="posts-index">
-                    {{ knot.name }}
-                </div>
-            </div>
-        </div>   
-    {% endfor %}
-{% endblock %}
