+++ templates/base.html
@@ -34,7 +34,6 @@
 
     <script src="/static/scripts/libs/jQuery.min.js"></script>
     <script src="/static/scripts/libs/fontAwesome.js"></script>
-    <script src="/static/scripts/button_to_top.js"></script>
     {% if dev_anv %}
         <script src="/static/scripts/libs/vue.js"></script>
     {% else %}
@@ -167,7 +166,6 @@
     </div>
     <script src="/static/scripts/main.js"></script>
     <script src="/static/scripts/libs/dateFormatter.js"></script>
-    <script src="/static/scripts/category_move.js"></script>
     <script src="/static/scripts/containers/base.js"></script>
     <script>
         $("div#alert-msg").fadeOut(4000);
