+++ templates/chat.html
@@ -61,7 +61,6 @@
     {% endif %}
 
     <script src="../static/scripts/socketio4.js" crossorigin="anonymous"></script>
-    <script src="../static/scripts/playNotification.js" type="text/javascript" ></script>
     <script src="../static/scripts/send_chat_msg.js" crossorigin="anonymous"></script>
 
 {% endblock %}
