+++ templates/edit_post.html
@@ -52,8 +52,7 @@
         {{ ckeditor.load() }}
         {{ ckeditor.config(name='body') }}
 
-        <script src="/static/scripts/jquery.js" type="text/javascript" charset="utf-8"></script>
-        <script src="/static/scripts/validate_form_input_length.js" type="text/javascript" charset="utf-8"></script>
+        <script src="/static/scripts/jquery.js"></script>
 
 	</div>
