+++ templates/emails/post_to_email.html
@@ -13,7 +13,7 @@
 {% endblock %}
 
 {% block link %}
-  <a href="{{ request.url_root + 'blog/' + post.slug }}" class="button button--blue">Go to post</a>
+  <a href="{{ request.url_root + 'blog/' + post.id }}" class="button button--blue">Go to post</a>
 {% endblock %}
 
 {% block safe_link %}
