+++ utils.py
@@ -1,3 +1,4 @@
+import re
 import hashlib
 import random
 import string
@@ -36,3 +37,8 @@ def args_to_parser(validator, params, location):
 
 def encrypt_string(string):
     return hashlib.sha3_256(string.encode()).hexdigest()
+
+
+def slugify(s):
+    pattern = r'[^\w+]'
+    return re.sub(pattern, '-', s)
